refactor: split pipeline.rs into submodules (#129)
Split the 1,677-line pipeline.rs into pipeline/mod.rs + four submodules:
- pipeline/state.rs: PipelineState, to_json, from_json
- pipeline/summary.rs: PipelineOutcome, TerminalReason, CapHitKind,
  IterationCounters, RunSummary, build_summary_artifact, iso8601 helpers
- pipeline/prompt.rs: SYSTEM_PREAMBLE, resolve_all_prompts,
  resolve_stage_prompt_inplace, inject_input, inject_note_block,
  prepend_preamble, read_prompt_file
- pipeline/routing.rs: StageOutcome, ExitKind, LoopControl, LoopOutcome,
  RouteTarget, PipelineProgress, run_stage, run_loop, route_pass,
  route_fail, build_name_index, handle_loop_outcome

Public interface from capsule::pipeline::* unchanged.
All 178 unit tests and integration tests pass.
