+++ src/pipeline/mod.rs
@@ -1,282 +1,33 @@
+mod prompt;
+mod routing;
+mod state;
+mod summary;
+
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
-use crate::config::{LoopConfig, OnFail, OnPass, PipelineConfig, PipelineEntry, StageConfig};
-use crate::verdict::{Verdict, VerdictStatus};
-use anyhow::{anyhow, Context};
+use crate::config::{PipelineConfig, PipelineEntry};
 
-pub const SYSTEM_PREAMBLE: &str = include_str!("../base-image/system_preamble.md");
+pub use prompt::SYSTEM_PREAMBLE;
+pub use state::PipelineState;
+pub use summary::{
+    build_summary_artifact, CapHitKind, IterationCounters, PipelineOutcome, RunSummary,
+    TerminalReason,
+};
 
-fn prepend_preamble(user_prompt: &str) -> String {
-    format!("{SYSTEM_PREAMBLE}\n\n{user_prompt}")
-}
-
-fn read_prompt_file(capsule_dir: &Path, path_str: &str) -> anyhow::Result<String> {
-    let path = capsule_dir.join(path_str);
-    let bytes = std::fs::read(&path)
-        .with_context(|| format!("prompt file not found: {}", path.display()))?;
-    Ok(String::from_utf8_lossy(&bytes).into_owned())
-}
+use routing::{
+    build_name_index, handle_loop_outcome, run_loop, run_stage, ExitKind, LoopControl,
+    PipelineProgress, StageOutcome,
+};
+use summary::CapHitKind as CapHit;
 
 pub trait StageRunner {
-    fn run(&mut self, stage_name: &str, prompt: &str, model: Option<&str>) -> Option<Verdict>;
-}
-
-/// High-level outcome of a pipeline run.
-#[derive(Debug, Clone, PartialEq)]
-pub enum PipelineOutcome {
-    /// All top-level stages completed, or a `done` verdict was emitted outside a loop.
-    Done,
-    /// An `exit` route was triggered; `from_fail` distinguishes fail-route vs pass-route.
-    Exit { from_fail: bool },
-    /// An iteration cap was exceeded.
-    CapHit,
-}
-
-/// Human-readable terminal reason written to the summary artifact.
-#[derive(Debug, Clone, PartialEq)]
-pub enum TerminalReason {
-    /// Pipeline completed normally (done verdict or all stages passed).
-    Done,
-    /// Exit via a pass-route (`on_pass: exit`).
-    Exit,
-    /// Exit via a fail-route (`on_fail: exit` or max_retries exceeded).
-    FailExit,
-    /// Iteration cap hit in a multi-stage pipeline.
-    CapHit,
-    /// Flat-form iteration limit reached (expected terminal for flat configs).
-    Ok,
-}
-
-/// Identifies which counter tripped when `CapHit` is the terminal reason.
-#[derive(Debug, Clone, PartialEq)]
-pub enum CapHitKind {
-    /// A loop's `max_iteration` was exceeded; `loop_idx` is the entry index.
-    LoopMaxIteration(usize),
-    /// The global `max_pipeline_iterations` was exceeded.
-    MaxPipelineIterations,
-}
-
-/// Snapshot of iteration counters at pipeline exit.
-#[derive(Debug, Clone, PartialEq)]
-pub struct IterationCounters {
-    pub global: u32,
-    /// Per-loop iteration counts, keyed by the loop's index in the top-level entries list.
-    pub loops: HashMap<usize, u32>,
-}
-
-/// Runtime state snapshot used to resume an interrupted pipeline.
-#[derive(Debug, Clone, PartialEq)]
-pub struct PipelineState {
-    pub current_idx: usize,
-    pub global_counter: u32,
-    pub fail_counts: HashMap<String, u32>,
-    pub last_stage: Option<String>,
-    pub last_verdict: Option<crate::verdict::Verdict>,
-    pub loop_iterations: HashMap<usize, u32>,
-    /// Run environment pairs persisted for resume; omitted from disk on successful exits.
-    pub env: Vec<(String, String)>,
-}
-
-impl PipelineState {
-    pub fn to_json(&self) -> serde_json::Value {
-        let fail_counts: serde_json::Map<String, serde_json::Value> = self
-            .fail_counts
-            .iter()
-            .map(|(k, v)| (k.clone(), serde_json::json!(v)))
-            .collect();
-        let loop_iterations: serde_json::Map<String, serde_json::Value> = self
-            .loop_iterations
-            .iter()
-            .map(|(k, v)| (k.to_string(), serde_json::json!(v)))
-            .collect();
-        let last_verdict = self
-            .last_verdict
-            .as_ref()
-            .map(|v| serde_json::to_value(v).unwrap_or(serde_json::Value::Null));
-        let env: serde_json::Map<String, serde_json::Value> = self
-            .env
-            .iter()
-            .map(|(k, v)| (k.clone(), serde_json::json!(v)))
-            .collect();
-        serde_json::json!({
-            "current_idx": self.current_idx,
-            "global_counter": self.global_counter,
-            "fail_counts": fail_counts,
-            "last_stage": self.last_stage,
-            "last_verdict": last_verdict,
-            "loop_iterations": loop_iterations,
-            "env": env,
-        })
-    }
-
-    pub fn from_json(v: &serde_json::Value) -> anyhow::Result<Self> {
-        let current_idx = v["current_idx"]
-            .as_u64()
-            .ok_or_else(|| anyhow!("pipeline_state.current_idx missing or invalid"))?
-            as usize;
-        let global_counter = v["global_counter"]
-            .as_u64()
-            .ok_or_else(|| anyhow!("pipeline_state.global_counter missing or invalid"))?
-            as u32;
-        let fail_counts: HashMap<String, u32> = v["fail_counts"]
-            .as_object()
-            .ok_or_else(|| anyhow!("pipeline_state.fail_counts missing or not an object"))?
-            .iter()
-            .map(|(k, val)| {
-                val.as_u64()
-                    .ok_or_else(|| anyhow!("pipeline_state.fail_counts[{}] is not a number", k))
-                    .map(|n| (k.clone(), n as u32))
-            })
-            .collect::<anyhow::Result<_>>()?;
-        let last_stage = v["last_stage"].as_str().map(str::to_owned);
-        let last_verdict: Option<crate::verdict::Verdict> = if v["last_verdict"].is_null() {
-            None
-        } else {
-            Some(
-                serde_json::from_value(v["last_verdict"].clone())
-                    .context("pipeline_state.last_verdict is malformed")?,
-            )
-        };
-        let loop_iterations: HashMap<usize, u32> = v["loop_iterations"]
-            .as_object()
-            .ok_or_else(|| anyhow!("pipeline_state.loop_iterations missing or not an object"))?
-            .iter()
-            .map(|(k, val)| {
-                let ki = k.parse::<usize>().map_err(|_| {
-                    anyhow!(
-                        "pipeline_state.loop_iterations key {:?} is not a valid index",
-                        k
-                    )
-                })?;
-                let vi = val.as_u64().ok_or_else(|| {
-                    anyhow!("pipeline_state.loop_iterations[{}] is not a number", k)
-                })? as u32;
-                Ok((ki, vi))
-            })
-            .collect::<anyhow::Result<_>>()?;
-        // env defaults to empty for backward compat (pre-ADR-0006 files lack this field)
-        let env: Vec<(String, String)> = v["env"]
-            .as_object()
-            .map(|obj| {
-                obj.iter()
-                    .map(|(k, val)| {
-                        let s = val
-                            .as_str()
-                            .ok_or_else(|| anyhow!("pipeline_state.env[{}] is not a string", k))?;
-                        Ok((k.clone(), s.to_owned()))
-                    })
-                    .collect::<anyhow::Result<_>>()
-            })
-            .transpose()?
-            .unwrap_or_default();
-        Ok(Self {
-            current_idx,
-            global_counter,
-            fail_counts,
-            last_stage,
-            last_verdict,
-            loop_iterations,
-            env,
-        })
-    }
-}
-
-/// Build the summary artifact JSON written to `last-run.json`.
-pub fn build_summary_artifact(
-    summary: &RunSummary,
-    workspace_dirty: bool,
-    pipeline_state: Option<&PipelineState>,
-) -> serde_json::Value {
-    let terminal_reason = match summary.terminal_reason {
-        TerminalReason::Done => "done",
-        TerminalReason::Exit => "exit",
-        TerminalReason::FailExit => "fail-exit",
-        TerminalReason::CapHit => "cap-hit",
-        TerminalReason::Ok => "ok",
-    };
-    let cap_hit_counter = match &summary.cap_hit {
-        None => serde_json::Value::Null,
-        Some(CapHitKind::LoopMaxIteration(idx)) => serde_json::json!({
-            "type": "max_iteration",
-            "loop_idx": idx,
-        }),
-        Some(CapHitKind::MaxPipelineIterations) => serde_json::json!({
-            "type": "max_pipeline_iterations",
-        }),
-    };
-    let last_verdict = summary
-        .last_verdict
-        .as_ref()
-        .map(|v| serde_json::to_value(v).unwrap_or(serde_json::Value::Null));
-    let loops: serde_json::Map<String, serde_json::Value> = summary
-        .iteration_counters
-        .loops
-        .iter()
-        .map(|(k, v)| (k.to_string(), serde_json::json!(v)))
-        .collect();
-    let ps = pipeline_state.map(|s| s.to_json());
-    serde_json::json!({
-        "terminal_reason": terminal_reason,
-        "cap_hit_counter": cap_hit_counter,
-        "last_stage": summary.last_stage,
-        "last_verdict": last_verdict,
-        "session_id": summary.session_id,
-        "iteration_counters": {
-            "global": summary.iteration_counters.global,
-            "loops": loops,
-        },
-        "pipeline_state": ps,
-        "timestamp": iso8601_now(),
-        "workspace_dirty": workspace_dirty,
-    })
-}
-
-fn iso8601_now() -> String {
-    use std::time::SystemTime;
-    let secs = SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs();
-    iso8601_from_secs(secs)
-}
-
-fn iso8601_from_secs(secs: u64) -> String {
-    let z = secs as i64 / 86400 + 719468;
-    let era = if z >= 0 { z } else { z - 146096 } / 146097;
-    let doe = z - era * 146097;
-    let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
-    let y = yoe + era * 400;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let d = doy - (153 * mp + 2) / 5 + 1;
-    let m = if mp < 10 { mp + 3 } else { mp - 9 };
-    let y = if m <= 2 { y + 1 } else { y };
-    let h = (secs / 3600) % 24;
-    let min = (secs / 60) % 60;
-    let s = secs % 60;
-    format!("{y:04}-{m:02}-{d:02}T{h:02}:{min:02}:{s:02}Z")
-}
-
-/// Mutable progress tracked across stage invocations during a pipeline run.
-struct PipelineProgress {
-    fail_counts: HashMap<String, u32>,
-    global_counter: u32,
-    input: Option<String>,
-    last_stage: Option<String>,
-    last_verdict: Option<Verdict>,
-}
-
-/// Execution summary produced by every `PipelineExecutor::run` call.
-#[derive(Debug, Clone, PartialEq)]
-pub struct RunSummary {
-    pub terminal_reason: TerminalReason,
-    pub last_stage: Option<String>,
-    pub last_verdict: Option<Verdict>,
-    pub iteration_counters: IterationCounters,
-    /// Which counter tripped, if `terminal_reason` is `CapHit` or `Ok`.
-    pub cap_hit: Option<CapHitKind>,
-    pub session_id: Option<String>,
+    fn run(
+        &mut self,
+        stage_name: &str,
+        prompt: &str,
+        model: Option<&str>,
+    ) -> Option<crate::verdict::Verdict>;
 }
 
 /// Return value of `PipelineExecutor::run`.
@@ -293,39 +44,6 @@ pub struct PipelineExecutor<R> {
     input: Option<String>,
     initial_state: Option<PipelineState>,
     capsule_dir: Option<PathBuf>,
-}
-
-fn resolve_all_prompts(
-    entries: &mut [PipelineEntry],
-    capsule_dir: Option<&Path>,
-) -> anyhow::Result<()> {
-    for entry in entries.iter_mut() {
-        match entry {
-            PipelineEntry::Stage(stage) => resolve_stage_prompt_inplace(stage, capsule_dir)?,
-            PipelineEntry::Loop(loop_config) => {
-                for stage in &mut loop_config.stages {
-                    resolve_stage_prompt_inplace(stage, capsule_dir)?;
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-fn resolve_stage_prompt_inplace(
-    stage: &mut StageConfig,
-    capsule_dir: Option<&Path>,
-) -> anyhow::Result<()> {
-    stage.prompt = match (stage.prompt.as_deref(), capsule_dir) {
-        (Some(path_str), Some(dir)) => {
-            let content = read_prompt_file(dir, path_str)
-                .with_context(|| format!("stage '{}': failed to load prompt", stage.name))?;
-            Some(prepend_preamble(&content))
-        }
-        (Some(literal), None) => Some(literal.to_string()),
-        (None, _) => None,
-    };
-    Ok(())
 }
 
 impl<R: StageRunner> PipelineExecutor<R> {
@@ -360,7 +78,7 @@ impl<R: StageRunner> PipelineExecutor<R> {
     }
 
     pub fn run(mut self) -> anyhow::Result<PipelineRunResult> {
-        resolve_all_prompts(&mut self.config.entries, self.capsule_dir.as_deref())?;
+        prompt::resolve_all_prompts(&mut self.config.entries, self.capsule_dir.as_deref())?;
 
         let name_to_entry = build_name_index(&self.config);
         let max_pipeline = self.config.max_pipeline_iterations;
@@ -399,10 +117,7 @@ impl<R: StageRunner> PipelineExecutor<R> {
             match &self.config.entries[current_idx] {
                 PipelineEntry::Stage(stage) => {
                     if progress.global_counter >= max_pipeline {
-                        break (
-                            PipelineOutcome::CapHit,
-                            Some(CapHitKind::MaxPipelineIterations),
-                        );
+                        break (PipelineOutcome::CapHit, Some(CapHit::MaxPipelineIterations));
                     }
                     progress.global_counter += 1;
                     match run_stage(&mut self.runner, stage, &name_to_entry, &mut progress)? {
@@ -492,361 +207,6 @@ impl<R: StageRunner> PipelineExecutor<R> {
             },
             pipeline_state,
         })
-    }
-}
-
-#[derive(Clone)]
-enum ExitKind {
-    PassRoute,
-    FailRoute,
-}
-
-enum LoopCapKind {
-    MaxIteration,
-    MaxPipelineIterations,
-}
-
-enum StageOutcome {
-    Advance(usize),
-    AdvanceIntoLoop { entry_idx: usize, stage_idx: usize },
-    Done,
-    Exit(ExitKind),
-}
-
-enum RouteTarget {
-    Entry(usize),
-    LoopStage { entry_idx: usize, stage_idx: usize },
-}
-
-enum LoopControl {
-    Advance(usize),
-    Break(PipelineOutcome, Option<CapHitKind>),
-}
-
-enum LoopOutcome {
-    LoopDone { iterations: u32 },
-    Exit { kind: ExitKind, iterations: u32 },
-    CapHit { kind: LoopCapKind, iterations: u32 },
-}
-
-fn run_loop(
-    runner: &mut dyn StageRunner,
-    loop_config: &LoopConfig,
-    progress: &mut PipelineProgress,
-    max_pipeline: u32,
-    start_stage_idx: usize,
-) -> anyhow::Result<LoopOutcome> {
-    let loop_name_to_idx: HashMap<String, usize> = loop_config
-        .stages
-        .iter()
-        .enumerate()
-        .map(|(i, s)| (s.name.clone(), i))
-        .collect();
-
-    let mut iteration_count: u32 = 0;
-    let mut stage_idx: usize = start_stage_idx;
-    let mut retrying_top = false;
-
-    loop {
-        if stage_idx >= loop_config.stages.len() {
-            stage_idx = 0;
-            retrying_top = false;
-        }
-
-        if stage_idx == 0 && !retrying_top {
-            iteration_count += 1;
-            if let Some(max) = loop_config.max_iteration {
-                if iteration_count > max {
-                    return Ok(LoopOutcome::CapHit {
-                        kind: LoopCapKind::MaxIteration,
-                        iterations: iteration_count - 1,
-                    });
-                }
-            }
-        }
-        retrying_top = false;
-
-        if progress.global_counter >= max_pipeline {
-            return Ok(LoopOutcome::CapHit {
-                kind: LoopCapKind::MaxPipelineIterations,
-                iterations: iteration_count,
-            });
-        }
-        progress.global_counter += 1;
-
-        let stage = &loop_config.stages[stage_idx];
-        let base_prompt = stage.prompt.as_deref().unwrap_or("");
-        let with_input = inject_input(&mut progress.input, base_prompt);
-        let effective_prompt = inject_note_block(
-            progress.last_stage.as_deref(),
-            &progress.last_verdict,
-            &with_input,
-        );
-        progress.last_stage = Some(stage.name.clone());
-        let verdict = runner.run(&stage.name, &effective_prompt, stage.model.as_deref());
-        progress.last_verdict = verdict.clone();
-
-        if matches!(
-            verdict.as_ref().map(|v| &v.status),
-            Some(VerdictStatus::Done)
-        ) {
-            return Ok(LoopOutcome::LoopDone {
-                iterations: iteration_count,
-            });
-        }
-
-        let is_pass = matches!(
-            verdict.as_ref().map(|v| &v.status),
-            Some(VerdictStatus::Pass)
-        );
-
-        if is_pass {
-            progress.fail_counts.remove(&stage.name);
-            match &stage.on_pass {
-                OnPass::Next => stage_idx += 1,
-                OnPass::Stage(name) => match loop_name_to_idx.get(name.as_str()) {
-                    Some(&idx) => stage_idx = idx,
-                    None => {
-                        return Ok(LoopOutcome::Exit {
-                            kind: ExitKind::PassRoute,
-                            iterations: iteration_count,
-                        })
-                    }
-                },
-                OnPass::Exit => {
-                    return Ok(LoopOutcome::Exit {
-                        kind: ExitKind::PassRoute,
-                        iterations: iteration_count,
-                    })
-                }
-            }
-        } else {
-            let fail_count = progress.fail_counts.entry(stage.name.clone()).or_insert(0);
-            *fail_count += 1;
-
-            if let Some(max) = stage.max_retries {
-                if *fail_count > max {
-                    return Ok(LoopOutcome::Exit {
-                        kind: ExitKind::FailRoute,
-                        iterations: iteration_count,
-                    });
-                }
-            }
-
-            match &stage.on_fail {
-                OnFail::Exit => {
-                    return Ok(LoopOutcome::Exit {
-                        kind: ExitKind::FailRoute,
-                        iterations: iteration_count,
-                    })
-                }
-                OnFail::Retry => {
-                    if stage_idx == 0 {
-                        retrying_top = true;
-                    }
-                }
-                OnFail::Stage(name) => match loop_name_to_idx.get(name.as_str()) {
-                    Some(&idx) => stage_idx = idx,
-                    None => {
-                        return Ok(LoopOutcome::Exit {
-                            kind: ExitKind::FailRoute,
-                            iterations: iteration_count,
-                        })
-                    }
-                },
-            }
-        }
-    }
-}
-
-/// Builds a map from stage name to its resolved route target.
-/// Top-level stages map to `Entry(i)`; loop-internal stages map to `LoopStage`.
-fn build_name_index(config: &PipelineConfig) -> HashMap<String, RouteTarget> {
-    let mut map = HashMap::new();
-    for (i, entry) in config.entries.iter().enumerate() {
-        match entry {
-            PipelineEntry::Stage(s) => {
-                map.insert(s.name.clone(), RouteTarget::Entry(i));
-            }
-            PipelineEntry::Loop(l) => {
-                for (j, s) in l.stages.iter().enumerate() {
-                    map.insert(
-                        s.name.clone(),
-                        RouteTarget::LoopStage {
-                            entry_idx: i,
-                            stage_idx: j,
-                        },
-                    );
-                }
-            }
-        }
-    }
-    map
-}
-
-fn handle_loop_outcome(
-    outcome: LoopOutcome,
-    entry_idx: usize,
-    loop_iterations: &mut HashMap<usize, u32>,
-) -> LoopControl {
-    match outcome {
-        LoopOutcome::LoopDone { iterations } => {
-            loop_iterations.insert(entry_idx, iterations);
-            LoopControl::Advance(entry_idx + 1)
-        }
-        LoopOutcome::Exit { kind, iterations } => {
-            loop_iterations.insert(entry_idx, iterations);
-            LoopControl::Break(
-                PipelineOutcome::Exit {
-                    from_fail: matches!(kind, ExitKind::FailRoute),
-                },
-                None,
-            )
-        }
-        LoopOutcome::CapHit {
-            kind: LoopCapKind::MaxIteration,
-            iterations,
-        } => {
-            loop_iterations.insert(entry_idx, iterations);
-            LoopControl::Break(
-                PipelineOutcome::CapHit,
-                Some(CapHitKind::LoopMaxIteration(entry_idx)),
-            )
-        }
-        LoopOutcome::CapHit {
-            kind: LoopCapKind::MaxPipelineIterations,
-            iterations,
-        } => {
-            loop_iterations.insert(entry_idx, iterations);
-            LoopControl::Break(
-                PipelineOutcome::CapHit,
-                Some(CapHitKind::MaxPipelineIterations),
-            )
-        }
-    }
-}
-
-fn inject_input(input: &mut Option<String>, base_prompt: &str) -> String {
-    if let Some(text) = input.take() {
-        format!("<capsule:input>\n{text}\n</capsule:input>\n\n{base_prompt}")
-    } else {
-        base_prompt.to_string()
-    }
-}
-
-fn inject_note_block(
-    last_stage: Option<&str>,
-    last_verdict: &Option<Verdict>,
-    base_prompt: &str,
-) -> String {
-    if let (Some(name), Some(verdict)) = (last_stage, last_verdict.as_ref()) {
-        let notes = verdict.notes.as_deref().filter(|n| !n.is_empty());
-        if let Some(notes) = notes {
-            let status_str = match verdict.status {
-                VerdictStatus::Pass => "pass",
-                VerdictStatus::Fail => "fail",
-                VerdictStatus::Done => "done",
-            };
-            let block = format!(
-                "<previous-stage>\nStage: {name}\nStatus: {status_str}\nNotes: {notes}\n</previous-stage>"
-            );
-            return format!("{block}\n\n{base_prompt}");
-        }
-    }
-    base_prompt.to_string()
-}
-
-fn run_stage(
-    runner: &mut dyn StageRunner,
-    stage: &StageConfig,
-    name_to_entry: &HashMap<String, RouteTarget>,
-    progress: &mut PipelineProgress,
-) -> anyhow::Result<StageOutcome> {
-    let base_prompt = stage.prompt.as_deref().unwrap_or("");
-    let with_input = inject_input(&mut progress.input, base_prompt);
-    let effective_prompt = inject_note_block(
-        progress.last_stage.as_deref(),
-        &progress.last_verdict,
-        &with_input,
-    );
-    progress.last_stage = Some(stage.name.clone());
-    let verdict = runner.run(&stage.name, &effective_prompt, stage.model.as_deref());
-    progress.last_verdict = verdict.clone();
-
-    if matches!(
-        verdict.as_ref().map(|v| &v.status),
-        Some(VerdictStatus::Done)
-    ) {
-        return Ok(StageOutcome::Done);
-    }
-
-    let is_pass = matches!(
-        verdict.as_ref().map(|v| &v.status),
-        Some(VerdictStatus::Pass)
-    );
-
-    if is_pass {
-        progress.fail_counts.remove(&stage.name);
-        Ok(route_pass(stage, name_to_entry))
-    } else {
-        let fail_count = progress.fail_counts.entry(stage.name.clone()).or_insert(0);
-        *fail_count += 1;
-
-        if let Some(max) = stage.max_retries {
-            if *fail_count > max {
-                return Ok(StageOutcome::Exit(ExitKind::FailRoute));
-            }
-        }
-
-        Ok(route_fail(stage, name_to_entry))
-    }
-}
-
-fn route_pass(stage: &StageConfig, name_to_entry: &HashMap<String, RouteTarget>) -> StageOutcome {
-    match &stage.on_pass {
-        OnPass::Next => {
-            let idx = match name_to_entry.get(&stage.name) {
-                Some(RouteTarget::Entry(i)) => *i,
-                _ => 0,
-            };
-            StageOutcome::Advance(idx + 1)
-        }
-        OnPass::Stage(name) => match name_to_entry.get(name.as_str()) {
-            Some(RouteTarget::Entry(idx)) => StageOutcome::Advance(*idx),
-            Some(RouteTarget::LoopStage {
-                entry_idx,
-                stage_idx,
-            }) => StageOutcome::AdvanceIntoLoop {
-                entry_idx: *entry_idx,
-                stage_idx: *stage_idx,
-            },
-            None => StageOutcome::Exit(ExitKind::PassRoute),
-        },
-        OnPass::Exit => StageOutcome::Exit(ExitKind::PassRoute),
-    }
-}
-
-fn route_fail(stage: &StageConfig, name_to_entry: &HashMap<String, RouteTarget>) -> StageOutcome {
-    match &stage.on_fail {
-        OnFail::Exit => StageOutcome::Exit(ExitKind::FailRoute),
-        OnFail::Retry => {
-            let idx = match name_to_entry.get(&stage.name) {
-                Some(RouteTarget::Entry(i)) => *i,
-                _ => 0,
-            };
-            StageOutcome::Advance(idx)
-        }
-        OnFail::Stage(name) => match name_to_entry.get(name.as_str()) {
-            Some(RouteTarget::Entry(idx)) => StageOutcome::Advance(*idx),
-            Some(RouteTarget::LoopStage {
-                entry_idx,
-                stage_idx,
-            }) => StageOutcome::AdvanceIntoLoop {
-                entry_idx: *entry_idx,
-                stage_idx: *stage_idx,
-            },
-            None => StageOutcome::Exit(ExitKind::FailRoute),
-        },
     }
 }
 
@@ -1657,21 +1017,5 @@ mod tests {
         let mut v = full_state().to_json();
         v["env"]["KEY"] = serde_json::json!(42);
         assert!(PipelineState::from_json(&v).is_err());
-    }
-
-    #[test]
-    fn iso8601_epoch_zero() {
-        assert_eq!(iso8601_from_secs(0), "1970-01-01T00:00:00Z");
-    }
-
-    #[test]
-    fn iso8601_known_date() {
-        // 2025-01-01T00:00:00Z = 1735689600
-        assert_eq!(iso8601_from_secs(1_735_689_600), "2025-01-01T00:00:00Z");
-    }
-
-    #[test]
-    fn iso8601_mid_day() {
-        assert_eq!(iso8601_from_secs(961_073_130), "2000-06-15T12:45:30Z");
     }
 }
